move Write Review button up a level, handle zero reviews
