+++ client/src/components/RatingsReviews/Review.jsx
@@ -26,9 +26,9 @@ export default function Review(props) {
   return (
     <div id='Review' data-testid='Review' reviewid={props.review.review_id}>
       <StarRating rating={props.review.rating} />
-      <h3 data-testid='reviewsummary'>{props.review.summary}</h3>
       <h5 data-testid='reviewer_name'>{props.review.reviewer_name}</h5>
       <h5 data-testid='reviewdate'>{props.review.date}</h5>
+      <h3 data-testid='reviewsummary'>{props.review.summary}</h3>
       <p>{reviewBody}</p>
       {showMoreBtn}
       <p>Recommend: {props.review.recommend ? 'True' : 'False'}</p>
