+++ client/src/components/RatingsReviews/ReviewsList.css
@@ -1,0 +1,4 @@
+#ReviewScroll {
+  max-height: 85vh;
+  overflow-y: scroll;
+}
