+++ client/src/components/RatingsReviews/ReviewsList.jsx
@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from 'react';
+import './ReviewsList.css';
 
 import Review from './Review.jsx';
 
@@ -20,20 +21,18 @@ export const ReviewsList = function(props) {
   if (maxVisible < props.reviewslist.results.length) {
     showMoreBtn = (
       <button onClick={showMoreHandler} >More Reviews</button>
-    )
-
+    );
     displayAmount = maxVisible;
   } else {
     displayAmount = props.reviewslist.results.length;
   }
-  
   
   for (let i = 0; i < displayAmount ; i++) {
     displayedReviews.push(props.reviewslist.results[i]);
   }
   
   return (
-    <div id='ReviewsList' data-testid='ReviewsList'>
+    <div id='ReviewsList' data-testid='ReviewsList' >
       <h3>{props.reviewslist.results.length} reviews, sorted by 
         <select name='reviewsort' id='reviewsort' value={props.sort} onChange={(e) => {props.setSort(e.target.value)}} data-testid='select' >
           <option value={option.RELEVANT} key='relevant'>Relevant</option>
@@ -41,20 +40,22 @@ export const ReviewsList = function(props) {
           <option value={option.HELPFUL} key='helpful'>Helpful</option>
         </select>
       </h3>
-      {displayedReviews.map((reviewdata) => {
-        if(Object.keys(props.filter).length > 0) {
-          // there is a filter in place, so only let those reviews through
-          if(props.filter[reviewdata.rating]){
+      <div id='ReviewScroll'>
+        {props.reviewslist.results.length !== 0 ? displayedReviews.map((reviewdata) => {
+          if(Object.keys(props.filter).length > 0) {
+            // there is a filter in place, so only let those reviews through
+            if(props.filter[reviewdata.rating]){
+              return (
+                <Review review={reviewdata} key={reviewdata.review_id}/>
+              );
+            }
+          } else {
             return (
               <Review review={reviewdata} key={reviewdata.review_id}/>
             );
           }
-        } else {
-          return (
-            <Review review={reviewdata} key={reviewdata.review_id}/>
-          );
-        }
-      })}
+        }) : null}
+      </div>
       {showMoreBtn}
     </div>
   );
